Split Properties tests into multiple files.
The original property testing used a single file to specify all
property tests, which caused compilation issues and intermittent
Travis CI build failures.  This commit breaks the Properties file into
multiple files to make it faster to compile.

    | Properties  |   GHC | QuickCheck | Clean build/test      | properties build/test |
    |-------------+-------+------------+-----------------------+-----------------------|
    | single file | 8.4.4 |     2.11.3 | 3m48s / 5m56s / 1m33s | 1m42s / 1m58s / 0m31s |
    |             |       |   2.12.6.1 | 3m27s / 5m42s / 0m58s | 1m45s / 2m2s / 0m33s  |
    |-------------+-------+------------+-----------------------+-----------------------|
    | multi-file  | 8.4.4 |     2.11.3 | 2m19s / 4m26s / 0m37s | 0m35s / 0m44s / 0m9s  |
    |             |       |   2.12.6.1 | 2m20s / 4m26s / 0m37s | 0m36s / 0m45s / 0m10s |

(times are in real, user, and sys time)
